Refactor the code. move to another package
