+++ lang-ext/spring-boot-transaction/ballerina-txn-common/pom.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="UTF-8"?>
 <!--
   ~ Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
   ~
@@ -19,16 +18,16 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>declarative-transaction-sample</artifactId>
-    <packaging>jar</packaging>
-
     <parent>
-        <groupId>io.ballerina.springboot.sample</groupId>
-        <artifactId>spring-boot-samples</artifactId>
+        <groupId>io.ballerina.springboot</groupId>
+        <artifactId>spring-boot-transaction</artifactId>
         <version>1.0.0-SNAPSHOT</version>
     </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>ballerina-txn-common</artifactId>
+    <name>Ballerina - Spring Boot Transaction Core</name>
+    <packaging>jar</packaging>
 
     <dependencies>
         <dependency>
@@ -36,20 +35,20 @@
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
         <dependency>
+            <groupId>javax.transaction</groupId>
+            <artifactId>jta</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.springframework</groupId>
-            <artifactId>spring-jdbc</artifactId>
+            <artifactId>spring-tx</artifactId>
         </dependency>
         <dependency>
-            <groupId>mysql</groupId>
-            <artifactId>mysql-connector-java</artifactId>
+            <groupId>org.apache.httpcomponents</groupId>
+            <artifactId>httpclient</artifactId>
         </dependency>
         <dependency>
-            <groupId>io.ballerina.springboot</groupId>
-            <artifactId>spring-boot-common</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.ballerina.springboot</groupId>
-            <artifactId>spring-boot-jdbc</artifactId>
+            <groupId>commons-httpclient</groupId>
+            <artifactId>commons-httpclient</artifactId>
         </dependency>
     </dependencies>
 </project>
