+++ lang-ext/spring-boot-transaction/ballerina-txn-jdbc/pom.xml
@@ -19,13 +19,16 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>springboot-parent</artifactId>
         <groupId>io.ballerina.springboot</groupId>
+        <artifactId>spring-boot-transaction</artifactId>
         <version>1.0.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
+    <name>Ballerina - Spring Boot JDBC Transactions</name>
+    <url>http://ballerinalang.org</url>
 
-    <artifactId>spring-boot-jdbc</artifactId>
+    <artifactId>ballerina-txn-jdbc</artifactId>
+    <packaging>jar</packaging>
 
     <dependencies>
         <dependency>
@@ -34,7 +37,7 @@
         </dependency>
         <dependency>
             <groupId>io.ballerina.springboot</groupId>
-            <artifactId>spring-boot-common</artifactId>
+            <artifactId>ballerina-txn-common</artifactId>
         </dependency>
     </dependencies>
 </project>
