--- samples/spring-boot-samples/declarative-transaction-sample/src/main/resources/data.sql
@@ -1,7 +1,0 @@
-INSERT INTO BOOKING(ID,FIRST_NAME) VALUES ('12345','Ballerina');
-INSERT INTO BOOKING(ID,FIRST_NAME) VALUES ('2s456','Trigger Stack');
-INSERT INTO BOOKING(ID,FIRST_NAME) VALUES ('2dse45','EI');
-
-INSERT INTO ROOMS(ID,ROOM_NO) VALUES ('12345',3);
-INSERT INTO ROOMS(ID,ROOM_NO) VALUES ('2s456',4);
-INSERT INTO ROOMS(ID,ROOM_NO) VALUES ('2dse45',5);
