--- samples/spring-boot-samples/programmatic-transaction-sample/src/main/java/io/ballerina/springboot/samples/programmatic/HotelReservationRes.java
@@ -1,21 +1,0 @@
-package io.ballerina.springboot.samples.programmatic;
-
-
-public class HotelReservationRes {
-
-    private final long id;
-    private final String name;
-
-    public HotelReservationRes(long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
