--- samples/spring-boot-samples/programmatic-transaction-sample/src/main/resources/application.properties
@@ -1,11 +1,0 @@
-# H2
-#spring.h2.console.enabled=true
-#spring.h2.console.path=/h2
-
-# Datasource
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://localhost:3306/hotel_db?useSSL=false
-spring.datasource.username=root
-spring.datasource.password=admin
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
