+++ samples/transactions/ballerina-springboot-transactions/hotel_service_springboot/src/main/java/ballerina/sidecar/samples/transactions/BookingService.java
@@ -1,40 +1,25 @@
-/*
-*  Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-*
-*  WSO2 Inc. licenses this file to you under the Apache License,
-*  Version 2.0 (the "License"); you may not use this file except
-*  in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-*  Unless required by applicable law or agreed to in writing,
-*  software distributed under the License is distributed on an
-*  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-*  KIND, either express or implied.  See the License for the
-*  specific language governing permissions and limitations
-*  under the License.
-*/
+package io.ballerina.springboot.samples.programmatic;
 
-package ballerina.sidecar.samples.transactions;
-
-
-import java.util.List;
-
+import io.ballerina.springboot.TransactionCallbackService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.boot.autoconfigure.*;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.TransactionDefinition;
 import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.DefaultTransactionDefinition;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+import javax.sql.DataSource;
+import java.util.List;
 
 import static org.springframework.web.bind.annotation.RequestMethod.GET;
 import static org.springframework.web.bind.annotation.RequestMethod.POST;
@@ -47,21 +32,34 @@ public class BookingService {
 
     private final static Logger logger = LoggerFactory.getLogger(BookingService.class);
 
-    private final JdbcTemplate jdbcTemplate;
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
+
+    @Autowired
+    private DataSource dataSource;
+
+    @Autowired
     private PlatformTransactionManager transactionManager;
 
-
-    public BookingService(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        transactionManager = new DataSourceTransactionManager(jdbcTemplate.getDataSource());
+    public BookingService() {
     }
 
     @RequestMapping(value = "/reservation/hotel", method = POST)
     public @ResponseBody
     HotelReservationRes reserveHotel(@RequestBody HotelReservationReq req) {
-        String name = req.getFullName();
-        reserveHotel(name);
-        return new HotelReservationRes(10001, "Hotel Reserved for " + name );
+        TransactionDefinition def = new DefaultTransactionDefinition();
+        TransactionStatus status = transactionManager.getTransaction(def);
+        try {
+            String name = req.getFullName();
+            reserveRoom(req.getRooms());
+            reserveHotel(name);
+            transactionManager.commit(status);
+            return new HotelReservationRes(10001, "Hotel Reserved for " + name);
+        } catch (DataAccessException dae) {
+            logger.error("Error in creating record, rolling back");
+            transactionManager.rollback(status);
+            throw dae;
+        }
     }
 
     @RequestMapping(value = "/reservation/hotel", method = GET)
@@ -75,9 +73,17 @@ public class BookingService {
         return response;
     }
 
+    @RequestMapping(value = "/reservation/room", method = GET)
+    @ResponseBody
+    public String getReservedRooms() {
+        String response = "";
+        logger.info("Get ALL reservation rooms");
+        for (int roomNo : findAllRooms()) {
+            response = response.concat(roomNo + "\n");
+        }
+        return response;
+    }
 
-    //Declarative Transactions
-    @Transactional
     void bookHotel(String... persons) {
         for (String person : persons) {
             logger.info("Booking " + person + " in a seat...");
@@ -85,23 +91,16 @@ public class BookingService {
         }
     }
 
-    // Programmatic Transaction Management
     void reserveHotel(String... persons) {
-
-        TransactionDefinition def = new DefaultTransactionDefinition();
-        TransactionStatus status = transactionManager.getTransaction(def);
-
-        try {
-            for (String person : persons) {
-                logger.info("Booking " + person + " in a room...");
-                jdbcTemplate.update("insert into BOOKINGS(FIRST_NAME) values (?)", person);
-            }
-            transactionManager.commit(status);
-        } catch (DataAccessException dae) {
-            logger.error("Error in creating record, rolling back");
-            transactionManager.rollback(status);
-            throw dae;
+        for (String person : persons) {
+            logger.info("Booking " + person + " in a room...");
+            jdbcTemplate.update("insert into BOOKINGS(FIRST_NAME) values (?)", person);
         }
+    }
+
+    void reserveRoom(int roomNo) {
+            logger.info("Booking room no: " + roomNo);
+            jdbcTemplate.update("insert into ROOMS(ROOM_NO) values (?)", roomNo);
     }
 
 
@@ -110,7 +109,12 @@ public class BookingService {
                 (rs, rowNum) -> rs.getString("FIRST_NAME"));
     }
 
+    List<Integer> findAllRooms() {
+        return jdbcTemplate.query("select ROOM_NO from ROOMS",
+                (rs, rowNum) -> rs.getInt("ROOM_NO"));
+    }
+
     public static void main(String[] args) {
-        SpringApplication.run(BookingService.class, args);
+        SpringApplication.run(new Class[]{BookingService.class, TransactionCallbackService.class}, args);
     }
 }
