+++ samples/transactions/ballerina-springboot-transactions/hotel_service_springboot/src/main/resources/schema.sql
@@ -1,2 +1,5 @@
 drop table BOOKINGS if exists;
-create table BOOKINGS(ID serial, FIRST_NAME varchar(10) NOT NULL);
+create table BOOKINGS(ID serial, FIRST_NAME varchar(20) NOT NULL);
+
+drop table ROOMS if exists;
+create table ROOMS (ID serial, ROOM_NO INT NOT NULL);
